settings arch refactor
