+++ lib/config/router/router.dart
@@ -1,6 +1,6 @@
 import 'package:go_router/go_router.dart';
 import 'package:todo_list/config/logger/navigation_logger.dart';
-import 'package:todo_list/features/settings/screen/settings_screen.dart';
+import 'package:todo_list/features/settings/presentation/screen/settings_screen.dart';
 import 'package:todo_list/features/todo/domain/entities/todo_entity.dart';
 import 'package:todo_list/features/todo/presentation/todo_all/screen/todo_all_screen.dart';
 import 'package:todo_list/features/todo/presentation/todo_single/screen/todo_single_screen.dart';
