+++ lib/core/services/settings/service/settings_service.dart
@@ -1,13 +1,13 @@
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:todo_list/core/services/settings/storage/settings_storage.dart';
-import 'package:todo_list/core/services/settings/storage/settings_storage_impl.dart';
-import '../setting/abstract/setting.dart';
+import 'package:todo_list/features/settings/domain/storage/settings_storage.dart';
+import 'package:todo_list/features/settings/data/settings_storage_impl.dart';
+import '../../../../features/settings/data/model/abstract/setting.dart';
 
-part '../setting/implementations/auth_setting.dart';
-part '../setting/implementations/delete_confirmation_setting.dart';
-part '../setting/implementations/locale_setting.dart';
-part '../setting/implementations/revision_setting.dart';
-part '../setting/implementations/theme_setting.dart';
+part '../../../../features/settings/data/model/implementations/auth_setting.dart';
+part '../../../../features/settings/data/model/implementations/delete_confirmation_setting.dart';
+part '../../../../features/settings/data/model/implementations/locale_setting.dart';
+part '../../../../features/settings/data/model/implementations/revision_setting.dart';
+part '../../../../features/settings/data/model/implementations/theme_setting.dart';
 
 class SettingsService {
   late LocaleSetting locale;
