+++ lib/features/settings/data/model/abstract/setting.dart
