+++ lib/features/settings/data/settings_storage_impl.dart
@@ -1,5 +1,5 @@
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:todo_list/core/services/settings/storage/settings_storage.dart';
+import 'package:todo_list/features/settings/domain/storage/settings_storage.dart';
 
 class SettingsStorageImpl implements SettingsStorage {
   final SharedPreferences _prefs;
