+++ lib/features/settings/domain/state_management/api_key/auth_state.dart
