+++ lib/features/settings/domain/state_management/auth/auth_state.dart
@@ -1,0 +1,8 @@
+part of 'auth_cubit.dart';
+
+final class AuthState {
+  final AuthSource source;
+  final String key;
+
+  AuthState({required this.source, required this.key});
+}
