+++ lib/features/settings/domain/state_management/auth/auth_type.dart
@@ -1,0 +1,3 @@
+part of 'auth_cubit.dart';
+
+enum AuthSource { env, manual }
