+++ lib/features/settings/domain/state_management/delete_confirmation/delete_confirmation_cubit.dart
