+++ lib/main.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:get_it/get_it.dart';
-import 'package:todo_list/core/services/settings/controllers/auth/auth_cubit.dart';
+import 'package:todo_list/features/settings/domain/state_management/auth/auth_cubit.dart';
 import 'package:todo_list/config/connectivity/connectivity_cubit.dart';
-import 'package:todo_list/core/services/settings/controllers/locale/locale_cubit.dart';
+import 'package:todo_list/features/settings/domain/state_management/locale/locale_cubit.dart';
 import 'package:todo_list/config/service_locator/service_locator.dart';
-import 'package:todo_list/core/services/settings/controllers/delete_confirmation/delete_confirmation_cubit.dart';
-import 'package:todo_list/core/services/settings/controllers/theme/theme_cubit.dart';
+import 'package:todo_list/features/settings/domain/state_management/delete_confirmation/delete_confirmation_cubit.dart';
+import 'package:todo_list/features/settings/domain/state_management/theme/theme_cubit.dart';
 import 'package:todo_list/features/todo/domain/repository/todo_repository.dart';
 import 'package:todo_list/features/todo/domain/state_management/todo_list_bloc/todo_list_bloc.dart';
 import 'package:todo_list/core/application/todo_app.dart';
